requirements.txt
cupy
base58
pynacl